tools/requirements.txt
selenium
python-dotenv
requests
pymupdf
google-generativeai
requests